Devided Role, Combine session online student with admin

+++ src/controller/admin/main/program/ProgMainDeleted.java
@@ -9,6 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import model.dao.ProgramDAO;
+import model.dto.UserDTO;
 
 /**
  * Servlet implementation class ProgMainDeleted
@@ -31,12 +32,14 @@ public class ProgMainDeleted extends HttpServlet {
 	 */
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try{
+			int user_level=((UserDTO)(request.getSession().getAttribute("adminsession"))).getUser_level();
+			
 			String branch_id=request.getParameter("branch_id");
 			String prog_id=request.getParameter("prog_id");
 			boolean check = false;
-			
-			check=new ProgramDAO().deleteProgMain(branch_id,prog_id);
-			
+			if(user_level==0){
+				check=new ProgramDAO().deleteProgMain(branch_id,prog_id);
+			}
 			if(check==true){
 				response.getWriter().write("true");
 			}else{
